+++ torch/onnx/_internal/fx/analysis/unsupported_nodes.py
@@ -56,10 +56,10 @@ class UnsupportedFxNodesAnalysis(_pass.Analysis):
             if node.op == "call_function":
                 try:
                     # NOTE: OPSchema matcher is not in this analysis scope.
-                    aten_name = self.dispatcher._get_aten_name(
+                    aten_name = self.onnxfunction_dispatcher.get_aten_name(
                         node, self.diagnostic_context
                     )
-                    self.dispatcher._get_function_overloads(
+                    self.onnxfunction_dispatcher.get_function_overloads(
                         node, aten_name, self.diagnostic_context
                     )
                 except diagnostics.RuntimeErrorWithDiagnostic as e:
